[Api][Admin] Enable Order resource
