+++ src/Sylius/Bundle/ApiBundle/Attribute/OrderTokenValueAware.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\ApiBundle\Attribute;
+
+#[\Attribute(\Attribute::TARGET_CLASS)]
+final class OrderTokenValueAware
+{
+    public function __construct(public string $constructorArgumentName = 'orderTokenValue')
+    {
+    }
+}
