+++ tests/Api/Responses/admin/order/get_order.json
@@ -9,7 +9,10 @@
         "@type": "Address",
         "firstName": "John",
         "lastName": "Doe",
+        "phoneNumber": null,
+        "company": null,
         "countryCode": "US",
+        "provinceName": null,
         "street": "Avenue",
         "city": "New York",
         "postcode": "90000"
@@ -19,7 +22,10 @@
         "@type": "Address",
         "firstName": "John",
         "lastName": "Doe",
+        "phoneNumber": null,
+        "company": null,
         "countryCode": "US",
+        "provinceName": null,
         "street": "Avenue",
         "city": "New York",
         "postcode": "90000"
@@ -48,6 +54,7 @@
     "tokenValue": "nAWw2jewpA",
     "id": "@integer@",
     "number": "@string@",
+    "notes": null,
     "items": [
         {
             "@id": "\/api\/v2\/admin\/order-items\/@integer@",
