--- tests/Api/Responses/admin/order/get_shipping_address_of_placed_order_response.json
@@ -1,16 +1,0 @@
-{
-    "@id": "\/api\/v2\/admin\/addresses\/@integer@",
-    "@context": "\/api\/v2\/contexts\/Address",
-    "@type": "Address",
-    "id": "@integer@",
-    "firstName": "Andrzej",
-    "lastName": "Legs",
-    "company": "Polmotors manufactory",
-    "street": "Moniuszki 16/20",
-    "countryCode": "PL",
-    "city": "Pabianice",
-    "postcode": "31-999",
-    "phoneNumber": "111999222",
-    "provinceCode": null,
-    "provinceName": "kujawsko-pomorskie"
-}
