--- tests/Api/Responses/admin/order/put_billing_address_of_placed_order_response.json
@@ -1,16 +1,0 @@
-{
-    "@id": "\/api\/v2\/admin\/addresses\/@integer@",
-    "@context": "\/api\/v2\/contexts\/Address",
-    "@type": "Address",
-    "id": "@integer@",
-    "firstName": "Updated: Adam",
-    "lastName": "Updated: Handley",
-    "company": "Updated: FMŻ",
-    "street": "Updated: Kościuszki 21",
-    "countryCode": "PL",
-    "city": "Updated: Bordeaux",
-    "postcode": "Updated: 99-999",
-    "phoneNumber": "Updated: 911213969",
-    "provinceCode": null,
-    "provinceName": "Updated: wielkopolskie"
-}
