--- tests/Api/Responses/admin/order/put_shipping_address_of_placed_order_response.json
@@ -1,16 +1,0 @@
-{
-    "@id": "\/api\/v2\/admin\/addresses\/@integer@",
-    "@context": "\/api\/v2\/contexts\/Address",
-    "@type": "Address",
-    "id": "@integer@",
-    "firstName": "Updated: Julia",
-    "lastName": "Updated: Kowalska",
-    "company": "Updated: Błysk",
-    "street": "Updated: Marszałkowska 10",
-    "countryCode": "GB",
-    "city": "Updated: Warszawa",
-    "postcode": "Updated: 00-001",
-    "phoneNumber": "Updated: 48222333444",
-    "provinceCode": null,
-    "provinceName": "Updated: mazowieckie"
-}
